--- FreeSql.DbContext/TempExtensions.cs
@@ -1,7 +1,0 @@
-﻿
-namespace FreeSql.Extensions.EntityUtil
-{
-    public static class TempExtensions
-    {
-    }
-}
